#14 [CHORE] 엔티티 파일들 entity 폴더로 이동
- 기능 패키지 밑에 있던 파일들을 /entity 폴더로 이동
