--- src/main/java/com/example/muaring/domain/member/InterestedMember.java
@@ -1,35 +1,0 @@
-package com.example.muaring.domain.member;
-
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(
-        name = "interested_member",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"follower_id", "followee_id"})
-        }
-)
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class InterestedMember {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "interested_member_id")
-    private Long id;
-
-    // 관심을 표시한 사용자
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "follower_id", nullable = false)
-    private Member follower;
-
-    // 관심을 받은 사용자
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "followee_id", nullable = false)
-    private Member followee;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-}
