--- src/main/java/com/example/muaring/domain/member/MemberGenrePreference.java
@@ -1,37 +1,0 @@
-package com.example.muaring.domain.member;
-
-import com.example.muaring.domain.music.Genre;
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(
-        name = "member_genre_preference",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"member_id", "genre_id"})
-        }
-)
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberGenrePreference {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_genre_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "genre_id", nullable = false)
-    private Genre genre;
-
-    @Column(nullable = false)
-    private Integer count;
-
-    @Column(name = "last_updated_at", nullable = false)
-    private LocalDateTime lastUpdatedAt;
-}
