+++ src/main/java/com/example/muaring/domain/member/entity/Follow.java
@@ -1,0 +1,35 @@
+package com.example.muaring.domain.member.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(
+        name = "follow",
+        uniqueConstraints = {
+                @UniqueConstraint(columnNames = {"follower_id", "followee_id"})
+        }
+)
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Follow {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "follow_id")
+    private Long id;
+
+    // 팔로우를 건 사용자
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "follower_id", nullable = false)
+    private Member follower;
+
+    // 팔로우 요청을 받은 사용자
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "followee_id", nullable = false)
+    private Member followee;
+
+    @Column(name = "created_at", nullable = false)
+    private LocalDateTime createdAt;
+}
