+++ src/main/java/com/example/muaring/domain/member/entity/Library.java
@@ -1,0 +1,34 @@
+package com.example.muaring.domain.member.entity;
+
+import com.example.muaring.domain.music.entity.Music;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+// 사용자별 라이브러리에 저장된 음악
+@Entity
+@Table(name = "library")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Library {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "library_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "music_id", nullable = false)
+    private Music music;
+
+    @Column(length = 20, nullable = false)
+    private String category;
+
+    @Column(name = "created_at", nullable = false)
+    private LocalDateTime createdAt;
+}
