+++ src/main/java/com/example/muaring/domain/member/entity/MemberMusicProfile.java
@@ -1,0 +1,73 @@
+package com.example.muaring.domain.member.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+import java.time.LocalDateTime;
+
+// 사용자 음악 취향 수치화한 테이블
+@Entity
+@Table(name = "member_music_profile")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MemberMusicProfile {
+
+    @Id
+    @Column(name = "member_id")
+    private Long id;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @MapsId
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @Column(name = "avg_danceability", nullable = false)
+    private Double avgDanceability;
+
+    @Column(name = "avg_energy", nullable = false)
+    private Double avgEnergy;
+
+    @Column(name = "avg_valence", nullable = false)
+    private Double avgValence;
+
+    @Column(name = "avg_acousticness", nullable = false)
+    private Double avgAcousticness;
+
+    @Column(name = "avg_instrumentalness", nullable = false)
+    private Double avgInstrumentalness;
+
+    @Column(name = "avg_speechiness", nullable = false)
+    private Double avgSpeechiness;
+
+    @Column(name = "avg_tempo", nullable = false)
+    private Double avgTempo;
+
+    @Column(name = "avg_loudness", nullable = false)
+    private Double avgLoudness;
+
+    @Column(name = "calculated_days", nullable = false)
+    private Integer calculatedDays;
+
+    @Column(name = "created_at", nullable = false)
+    private LocalDateTime createdAt;
+
+    @Column(name = "avg_popularity", nullable = false)
+    private Double avgPopularity;
+
+    @Column(name = "avg_rarity")
+    private Double avgRarity;
+
+    @Column(name = "hidden_gem_ratio", nullable = false)
+    private Double hiddenGemRatio;
+
+    @Column(name = "updated_at", nullable = false)
+    private LocalDateTime updatedAt;
+
+    @Enumerated(EnumType.STRING)
+    @Column(name = "status", nullable = false, length = 20)
+    private ProfileStatus status;
+
+    // 일단 한 파일에 넣었는데 분리 필요하면 분리하기
+    public enum ProfileStatus {
+        PROCESSING, COMPLETED
+    }
+}
