--- src/main/java/com/example/muaring/domain/notification/Notification.java
@@ -1,36 +1,0 @@
-package com.example.muaring.domain.notification;
-
-import com.example.muaring.domain.common.BaseEntity;
-import com.example.muaring.domain.member.Member;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "notification")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Notification extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "notification_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "receiver_id", nullable = false)
-    private Member receiver;
-
-    @Enumerated(EnumType.STRING)
-    @Column(length = 30, nullable = false)
-    private NotificationType type;
-
-    @Column(name = "target_id", nullable = false)
-    private Long targetId;
-
-    @Column(length = 255)
-    private String content;
-
-    @Column(name = "is_read", nullable = false)
-    private Boolean isRead = false;
-}
-
