--- src/main/java/com/example/muaring/domain/notification/NotificationType.java
@@ -1,11 +1,0 @@
-package com.example.muaring.domain.notification;
-
-public enum NotificationType {
-    MUSIC_CHECK_REMINDER,
-    MUSIC_UPLOAD_REMINDER,
-    LIKE,
-    COMMENT,
-    FOLLOW_APPROVED,
-    NEW_FOLLOWER,
-    INTERESTED_USER
-}
