+++ src/main/java/com/example/muaring/domain/notification/entity/NotificationType.java
@@ -1,0 +1,11 @@
+package com.example.muaring.domain.notification.entity;
+
+public enum NotificationType {
+    MUSIC_CHECK_REMINDER,
+    MUSIC_UPLOAD_REMINDER,
+    LIKE,
+    COMMENT,
+    FOLLOW_APPROVED,
+    NEW_FOLLOWER,
+    INTERESTED_USER
+}
