--- src/main/java/com/example/muaring/domain/social/Comment.java
@@ -1,34 +1,0 @@
-package com.example.muaring.domain.social;
-
-import com.example.muaring.domain.common.BaseEntity;
-import com.example.muaring.domain.member.Member;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Table(name = "comment")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    private MusicPost post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_comment_id")
-    private Comment parentComment;
-
-    @Column(length = 255, nullable = false)
-    private String content;
-
-}
