--- src/main/java/com/example/muaring/domain/social/Like.java
@@ -1,34 +1,0 @@
-package com.example.muaring.domain.social;
-
-import com.example.muaring.domain.member.Member;
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(
-        name = "`like`",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = {"post_id", "member_id"})
-        }
-)
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Like {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "like_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    private MusicPost post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-}
